Add tests for class Levels; add docs; add const qualifiers; change from Array to std::vector for field levels to avoid memory leaks.

+++ Dialogs.cpp
@@ -115,7 +115,7 @@ public:
   }
   void setCollection(int c, int levelInC)
   {
-    if (c < 0 || c >=m_levels->numCollections()) {
+    if (c < 0 || static_cast<unsigned int>(c) >=m_levels->numCollections()) {
       return;
     }    
     m_collection = c;
